Moved seeds_list model to base models.
