+++ source/apps/crawl_space/models.py
@@ -198,12 +198,3 @@ class Crawl(models.Model):
 
     def __unicode__(self):
         return self.name
-
-
-class SeedsList(models.Model):
-    name = models.CharField(max_length=64, unique=True,
-        validators=[alphanumeric_validator()])
-    seeds = models.TextField()
-
-    def __unicode__(self):
-        return self.name
